% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netint.R
\name{MS.UA.int}
\alias{MS.UA.int}
\title{"Memory Saving" Unweighted Average (UA) network integration}
\usage{
MS.UA.int(nets.files, example.names = NULL)
}
\arguments{
\item{nets.files}{a list with the names of the .rda files storing the
matrices representing the weighted adjacency matrices of the nets.}

\item{example.names}{a list with the names of the examples stored in each
net. If NULL (def.) it is assumed that the matrices have exactly the
same examples in the same order, otherwise the matrices are arranged to
be correctly aligned.}
}
\value{
the integrated matrix : the matrix resulting from UA.
}
\description{
It performs the unweighted average integration between networks:
\loadmathjax
\mjsdeqn{\bar{w}_{ij} = \frac{1}{n} \sum_{d = 1}^n w_{ij}^d}
The matrices are read from files and loaded one at time in memory.
}
